objectifying into board and gamekeeper, also deletes a file wrongly committed
